% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\docType{data}
\name{effort}
\alias{effort}
\title{Data of visual detections by Protected Species Observers (PSOs) for the Community Offshore Wind Project, 2023}
\format{
A tibble with 2255 rows and 82 variables:
\describe{
\item{PROJECT_NUMBER}{Number identifier of wind farm development project}
\item{VESSEL}{Name of vessel}
}
}
\source{
\url{https://www.fisheries.noaa.gov/action/incidental-take-authorization-community-offshore-wind-llc-marine-site-characterization}
}
\usage{
effort
}
\description{
Data of visual detections by Protected Species Observers (PSOs) for the Community Offshore Wind Project, 2023
}
\keyword{datasets}
