% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEMADMDEBPSO-package.R
\docType{package}
\name{PEMADMDEBPSO-package}
\alias{PEMADMDEBPSO}
\alias{PEMADMDEBPSO-package}
\title{PEMADMDEBPSO: Data Package with Community Offshore Wind PSO Data}
\description{
This is a data package containing data from the Community Offshore Wind Development Project (2023). This is data from the High Resolution Geophysical Survey Protected Species Observer Final Report 2023.
}
\author{
\strong{Maintainer}: Nicole Mucci \email{nicole.mucci@noaa.gov}

}
\keyword{internal}
